Bug 1200194 - Add UpdateCaretsHint to guide appearance changing. r=roc
This is the patch which fixed the bug.

When calling OnScrollPositionChanged in cursor mode, we want the
appearance of the caret to be preserved since the caret might be hidden
due to timeout. We should respect the old appearance of the caret.

Add a marionette test to ensure the caret does not appear due to
ScrollPositionChanged.