Bug 1191819 - part 1 - substitute MOZ_BZ2_CFLAGS as a list; r=mshal
We appear to be inconsistent about which CFLAGS variables get
substituted as lists (most things relating to widget libraries, for
instance) and which are just raw strings (most everything else).  This
patch is a first step towards making everything a list, which makes the
next patch much easier to write.  The other variables can be converted
as a followup bug.