Bug 1199434 - Send printer error messages to the parent process to report to the user. r=Mossop
The printing back-end used to be in charge of opening up an error dialog
when things go wrong with printing. Now we fire an event and let PrintUtils
do the work of showing the error message. This has the added bonus of making
the error messages work with e10s.