bug 1192353 - make HyperTextProxyAccessibleWrap inherit from
HyperTextAccessibleWrap r=davidb

This is rather unfortunate, AccessibleWrap itself wastes a fair amount of space
when it just stores a pointer to a proxy, and this makes it waste even more.
However this is rather necessary for now because we need to be able to downcast
classes such as ia2AccessibleText to one type that works both when the
accessible is pointing to a proxy and when it is not.  That means
HyperTextAccessibleWrap and HyperTextProxyAccessibleWrap need to have the same
layout.