Bug 1209603 patch 6 - Prepare to use a different meaning of mBits when cached style data pointer is null. r=heycam
We currently only use the style struct bits in mBits when the style
context has the relevant struct cached.  The bit being set indicates
whether or not the context owns the struct.

This patch conditions the necessary tests on a cached struct being
present so that we can use (for reset structs, i.e., those with
non-inherited properties) mBits to mean something different when the
cached storage is null.